# Special case empty objects and arrays