@lemire 2020-08-03
The gain is remarkable!