@jkeiser 2020-08-03
I'm checking the small westmere drop detected by CI, but I think this is ready.