@jkeiser 2020-08-04
I measure a 1% speedup on twitter.json on Windows.